tests: Added unit test for team use-cases

+++ api/src/application/use-cases/team/DeleteTeamUseCase.js
@@ -5,7 +5,15 @@ class DeleteTeamUseCase {
   }
 
   async execute(teamId, projectMembersInTeam) {
+    if (!teamId) throw new Error('teamId was not provided');
+    if (!Array.isArray(projectMembersInTeam)) {
+      throw new Error('projectMembersInTeam is not an array');
+    }
+
     const teamDeleted = await this.teamRepository.delete(teamId);
+    if (teamDeleted === 0) {
+      throw new Error('Something went wrong deleting the team');
+    }
 
     const teamMembersDeletedFromProjects =
       projectMembersInTeam.length > 0
@@ -15,6 +23,12 @@ class DeleteTeamUseCase {
             ),
           )
         : [];
+
+    if (teamMembersDeletedFromProjects.some((t) => t === 0)) {
+      throw new Error(
+        `Something went wrong deleting some the project member. Result db: ${teamMembersDeletedFromProjects}`,
+      );
+    }
 
     return { teamDeleted, teamMembersDeletedFromProjects };
   }
